+++ frontend/src/App.js
@@ -1,19 +1,19 @@
 import React from "react";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import Inicio from "./components/inicio";
-import LoginAdmin from "./components/admin/login_Admin";
-import LoginUssers from "./components/usuario/login_Ussers";
-import SolicitudLibros from "./components/usuario/SolicitudLibros";
-import AgregarUssers from "./components/admin/agregarUsuarios";
-import AgregarLibro from "./components/admin/agregar_Libro";
-import PaletaPrestamos from "./components/admin/Paleta_Prestamos";
-import NavbarAdmin from "./components/navbar/NavbarAdmin";
-import NavbarUsuarios from "./components/navbar/NavbarUsuarios";
-import InicioAdmin from "./components/admin/InicioAdmin";
-import InicioUsuario from "./components/usuario/InicioUsuario";
-import LibrosDisponibles from "./components/usuario/LibrosDisponibles";
-import BlogForm from "./components/admin/BlogForm"
-import "./components/css/navbar.css";
+import LoginAdmin from "./admin/login_Admin";
+import LoginUssers from "./usuario/login_Ussers";
+import SolicitudLibros from "./usuario/SolicitudLibros";
+import AgregarUssers from "./admin/agregarUsuarios";
+import AgregarLibro from "./admin/agregar_Libro";
+import PaletaPrestamos from "./admin/Paleta_Prestamos";
+import NavbarAdmin from "./navbar/NavbarAdmin";
+import NavbarUsuarios from "./navbar/NavbarUsuarios";
+import InicioAdmin from "./admin/InicioAdmin";
+import InicioUsuario from "./usuario/InicioUsuario";
+import LibrosDisponibles from "./usuario/LibrosDisponibles";
+import BlogForm from "./admin/BlogForm"
+import "./css/navbar.css";
 //import axios from 'axios';
 
 // // Components
