+++ frontend/src/admin/Paleta_Prestamos.js
