+++ frontend/src/css/sigCanvas.css
